Add tests for the size of the exponent
I was wondering how Decimal determines the size of
the exponent when performing division, so I added
some tests.  The tests hold true for addition and
subtraction, but they fail for multiplication and
division.  Data.Decimal states that "Binary
arithmetic results have an exponent equal to the
maximum of the exponents of the arguments", but
currently this is not true.